add Yml type

+++ tests/Howyi/RetortTest.php
@@ -8,12 +8,29 @@ class RetortTest extends \PHPUnit\Framework\TestCase
 {
     public function testSeal()
     {
+        self::configZip();
+        Retort::seal();
+        self::configYml();
+        Retort::seal();
+        self::removeConfig();
         Retort::seal();
         $this->assertTrue(true);
     }
 
     public function testHeat()
     {
+        self::configZip();
+        Retort::heat();
+        self::configYml();
+        Retort::heat();
+        self::removeConfig();
+        Retort::heat();
+        $this->assertTrue(true);
+    }
+
+    public static function configZip()
+    {
+        self::removeConfig();
         $config = [
             'type'        => 'zip',
             'name'        => 'Retort',
@@ -23,10 +40,26 @@ class RetortTest extends \PHPUnit\Framework\TestCase
             ],
         ];
         file_put_contents('rtrt.yml', Yaml::dump($config, 4, 2));
-        Retort::heat();
-        unlink('rtrt.yml');
-        Retort::heat();
+    }
+
+    public static function configYml()
+    {
+        self::removeConfig();
+        $config = [
+            'type'        => 'yml',
+            'name'        => 'Retort',
+            'directories' => [
+                'src',
+                'tests',
+            ],
+        ];
         file_put_contents('rtrt.yml', Yaml::dump($config, 4, 2));
-        $this->assertTrue(true);
+    }
+
+    public static function removeConfig()
+    {
+        if (file_exists('rtrt.yml')) {
+            unlink('rtrt.yml');
+        }
     }
 }
